refactoring - create static utils
